Remove poll for layout in PictureInPictureControllerTest
It's unclear, without a local repro, if the layout check is the root
cause of the flaky test, or if there's an underlying issue.  This CL
removes it, to see if other things start to flake without it.

I don't think that the layout check is valuable anymore, but we'll
find that out too if different tests start flaking because of it.

All of it is stable locally, though.

Change-Id: If73fcbd255d088cf8d81b644fa99b554523caa53
Bug: 1219007
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2956662
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891889}