Fix StatusAreaWidget View ownership
This Widget held on to std::unique_ptr<View>, making them effectively
set_owned_by_client(). This relied on ~StatusAreaWidget happening before
~Widget would have deleted them from the View hierarchy as `delete View`
removes the View from the hierarchy (does not double-delete).

Owning Views that are inside the View hierarchy is being disallowed and
set_owned_by_client() is being removed. This changes the ownership model
so that StatusAreaWidget holds on to pointers owned by the View
hierarchy and the View hierarchy gets cleared out in ~Widget instead of
implicitly through ~unique_ptr in ~StatusAreaWidget.

Destruction still happens roughly at the same point in time but there
are no unique_ptrs to Views that are really owned by the hierarchy, or
Views owned by the hierarchy that are really owned outside of it.

Bug: 1044687
Change-Id: I84426108e90d99c544777e134f89b2115a7e3ce3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2957093
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891907}