Add a client-side experiment for extending Variations Safe Mode.
Note that iOS and Android are excluded from this experiment.

This is the third change in a series of changes for setting up the
experiment in variations_field_trial_creator.cc. This experiment has
three groups: Control, WritePrefs, and SignalEarlyAndWritePrefs. Clients
in the control group have the status quo behavior. Clients in the
WritePrefs group synchronously write prefs to disk during startup. This
allows us to measure the time that this write takes. Clients in the
SignalEarlyAndWritePrefs group (a) signal that Chrome has not yet exited
cleanly before setting up field trials and (b) synchronously write prefs
to disk.

The motivation behind this experiment is that Variations Safe Mode is
not triggered for crashes that occur while setting up field trials or
shortly thereafter. The cause is twofold. First, we do not signal that
Chrome has not yet exited cleanly until some point after field trials
are set up. Second, we fail to persist two prefs relevant to determining
whether to run in safe mode if we crash while setting up field trials or
shortly thereafter: kStabilityExitedCleanly and kVariationsCrashCount.

To persist these prefs to disk, we add a new function:
CommitPendingWriteSynchronously(). An existing function,
CommitPendingWrite() cannot be used to write the prefs to disk during
startup because CommitPendingWrite() schedules a write by posting a
task. However, at the point in which we're setting up field trials,
Chrome has not yet started threads and is not yet running tasks.

For more information about Variations Safe Mode, please see this doc:
https://docs.google.com/document/d/17UN2pLSa5JZqk8f3LeYZIftXewxqcITotgalTrJvGSY/

Bug: b/184937096
Change-Id: Iea4c468faedb39cb7e1b5bf564eecc4e0efb69a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2898323
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/master@{#891910}