Fix flaky RenderThreadImplBrowserTests
The last attempt at fixing this (crrev.com/c/2954844) introduced
ChildProcess::FlushIOThreadForTesting() to flush the in-process
renderer's IO thread so that it could not post. However flushing the
underlying base::Thread was not enough, because it could still process
tasks after flushing, which kept the existing race condition around.

This CL changes FlushIOThreadForTesting() => StopIOThreadForTesting()
so that once we stop the IO thread, it can never process tasks again,
and therefore it cannot post any more tasks back to the browser's
main thread.

I was finally able to more concretely reproduce the test flake and I've
confirmed that this fix works.

R=danakj@chromium.org

Bug: 1126157
Change-Id: I257684e2a30a0bd3f83fdf410b162c7ce4ff5825
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2957188
Commit-Queue: Dominic Farolino <dom@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891911}