[segmentation_platform] Use execution dummy when no TFLite is available
The core implementation of the ModelExecutionManager,
ModelExecutionManagerImpl has a tight coupling to TFLite and supporting
libraries through the ModelExecutor and ModelHandler implementations.

However, it is currently not possible to include TFLite across all
platforms, and this is guarded using the GN variable
build_with_tflite_lib and its related BUILDFLAG(BUILD_WITH_TFLITE_LIB).

For any platform we need to at some point decide whether to use the
dummy or real implementation of the ModelExecutionManager, and this
CL provides a factory which uses the build flag to decide which one to
use.

To ensure good test coverage, the DummyModelExecutionManager and its
related test is compiled on all platforms, though the factory will never
return it.

It also cleans up some deps and public_deps in our internal BUILD.gn
targets related to our dependencies on //components/optimization_guide.

Bug: 1210648
Change-Id: Ib86f075d1d8b5fb390d5c97c3d82774591ac5553
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2956192
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891969}