Collect Arc.ImeCount metric when no ARC IME is installed.
To correctly know how many users are using ARC++ IME support.
Before this CL, the metrics is collected only when one or more Android
IMEs are installed on a device.
Also the metric will be recorded only when the number of the installed
Android IMEs changes with this CL.

Test: unit_tests
Bug: 1218867
Change-Id: I0732ae281b49fcb30ca1437e08922b51bd8cdf58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2955742
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891974}