Add Incognito clear browsing data dialog for desktop.
This CL adds the Incognito clear browsing data dialog that is shown to
the user when a clear browsing data flow is triggered from Incognito.

It renables the "Clear browsing data..." option in three-dots menu >
More Tools and shows the dialog only when the user either clicks on that
option or uses a shortcut option.

The dialog is missing the header art which will be added in later CLs.
The actual dialog looks like this go/incognito-cbd-dialog-without-art.
The expected dialog can be found in the crbug link.

Bug: 1214196
Change-Id: I08a0915ca45a17464a4684c45b7450bab7322296
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2940294
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892034}