[Autofill] Added ContentAutofillRouter skeleton.
This CL adds the ContentAutofillRouter class without any side effects.

A ContentAutofillRouter instance is owned by
ContentAutofillDriverFactory (1 per WebContents), and
ContentAutofillDrivers hold a weak reference to it.

The event handlers of ContentAutofillDriver pass the event to
ContentAutofillRouter, which passes them back to – for now – the same
ContentAutofillDriver.

Bug: 1187842
Change-Id: I8daaf179bb943f1b91e3bcda877ed33e6f2594b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2879511
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/master@{#892042}