[PartitionAlloc] Use fewer bucket sizes.
Currently, we have up to 8 buckets per order. This may be a bit too
high, use only 4. In particular, a couple things are clearly suboptimal
as of now:
- Poor cache locality in the first buckets of the thread cache and the
  central allocator: as most of the first buckets sizes are not multiple
  of the alignment, they are not "real" buckets. This means that the
  first few buckets in the thread cache are not tightly packed.
- Too many buckets: Many buckets are barely used.
- Large metadata overhead: sizeof(ThreadCache) is large, for instance.

Having more buckets has pros and cons:
- Pros:
  - Less wasted memory by rounding up sizes to the nearest bucket
  - Possibly better memory locality, as size classes are less likely to be
    shared across unrelated objects
- Cons:
  - Thread cache and allocator metadata are larger, since there are more buckets
  - Increased virtual address space and memory footprint: a single allocation is
    sufficient to create an entire slot span

Based on testing, this change seems to be improving memory footprint on
most benchmarks, but it's hard to be sure without running all
benchmarks.

This is likely only a first step, however finer grained tuning requires
more invasive changes, such as changing the requested size -> bucket
mapping, which is very performance sensitive.

Note about performance: This is expected to *regress* the allocated
object size reported by malloc(). This is because malloc() reports the
slot size rather than actually allocated object size for most
allocations, and this commit strictly increases this one. However, it
may increase or decrease private memory footprint, which is the metric
we care about.

Bug: 998048, 1217582
Change-Id: I7990922a50926587cb3befb1a2adcf986e8a5d35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2940293
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892052}