[Blink] Fix presubmit tests
https://chromium-review.googlesource.com/c/chromium/src/+/2874622
updated Blink presubmit tests that checked that base::Callback is
flagged as disallowed in Blink to instead check that
base::{Once, Repeating}Callback are disallowed. However, it turns out
that that test was conceptually wrong (presumably stale), as
base::{Once, Repeating}Callback *are* allowed in Blink:

https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/tools/blinkpy/presubmit/audit_non_blink_usage.py;l=114?q=audit_&ss=chromium

The reason that the existing test wasn't failing is because
base::Callback was not in that list (or perhaps had been removed from
it earlier as part of the transition to {Once, Repeating}Callback).

It's not clear why these tests weren't triggered as part of that CL,
but in any case, this CL fixes them by moving the tests to verify that
these constructs are *allowed* in Blink.

Bug: 1218891
Change-Id: I18ecf82fabe56b96bdfbabafcbbdd35e9cd70edf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2959891
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: Philip Jägenstedt <foolip@chromium.org>
Auto-Submit: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892061}