PowerMode: Handle OnTraceLogEnabled() before task runner available
When startup-tracing, TraceLog can be enabled before the arbiter's task
runner is set and a SequencedTaskRunnerHandle exists. In that case, we
can't yet add an observer to the arbiter.

Instead postpone this until the task runner is set. This requires some
additional locking because now the task runner becoming available can
race with the tracing session being disabled (on different threads).

Fixed: 1218640
Change-Id: I6d72e129224a08b69572e7cd15e80f26e04c383f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2956731
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892093}