Improve CompositorFrame validation
Expand on the validation of CompositorFrames in SurfaceAggregator to
check render pass structure is valid. In particular, check that render
passes have unique ids and that CompositorRenderPassDrawQuads refer to a
render pass that is drawn before the quad (which is required). This also
moves the validation into ResolvedFrameData to make it easier to test.

One behaviour change here is that CompositorFrames containing a render
pass cycle are now rejected as invalid outright. Before the
CompositorRenderPassDrawQuad that embedded a later render pass would
just fail to produce anything since the render pass wouldn't be ready
yet. There shouldn't be any code producing render pass cycles so it
should be safe to reject them outright.

Bug: 1194082
Change-Id: I6273de0eb627fabce3b50e1bf8ff8abb5cad0e75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2953905
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: Peter McNeeley <petermcneeley@google.com>
Cr-Commit-Position: refs/heads/master@{#892094}