dcheck-on-cros: Keep DCHECK at microphone_mute_notification_delegate_impl.cc:103 disabled on CrOS.
Chrome DCHECKs are disabled on CrOS during automated testing and rarely
enabled otherwise. This check failed when tested on an informational
builder, so it should be completely disabled. This change will let us
enable DCHECK testing on CrOS.

The check should be re-enabled when the underlying bug is fixed
(crbug.com/1217883).

See go/chrome-dcheck-on-cros and go/dcheck-cleanup-workflow for more
information.

Bug: chromium:1217883
Change-Id: Ib1c0bc6d2e30c9d9acdb6ce7f486294f6d19f416
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2952736
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892105}