request_initiator_origin_lock checks in SignedExchangeURLLoaderFactory.
This CL replicates the functionality of GetTrustworthyInitiator
underneath SignedExchangeURLLoaderFactory.  This is sufficient to
enable removal of GetTrustworthyInitiator in a follow-up CL.

In the long-term a mismatched `request_initiator_origin_lock` should
result in erroring out the resource request, but this is postponed to a
follow-up CL later.  We should pursue this if the DumpWithoutCrashing
call from the current CL doesn't result in a high volume of reports.

Bug: 1217825
Change-Id: I95f5d834d9097e13049a3ad022fe90fbdefe5aee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2923326
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892142}