Consolidate successful scan tests.
This CL reduces the number of ScanServiceTest unit tests by handling PDF
file types in the (HoldingSpace)Scan tests. This change also makes the
tests more robust to future scanned file types by failing if a new type
is not explicitly added to the file type map, which should prompt
developers to consider whether any special handling is necessary for new
types (as it is for PDFs).

Bug: 1192276
Change-Id: I23ae56aecebef8c61d92845b4a523955296b763c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2956899
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Commit-Queue: Colin Kincaid <ckincaid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892146}