sharing: implement share target ranking algorithm
This change implements the algorithm described in the "focus doc", which
will shortly be checked into the tree once it is rewritten into Markdown.
The algorithm decides which share targets the share sheet should present
to the user, taking into account existing positions of targets, which
are available on the system and able to handle shares of the current
type, recent usage, and all-time usage of each target.

This change also adds more test coverage for the ranking algorithm,
and enables all the existing disabled test coverage.

Note that the new algorithm is not yet used anywhere - a followup
CL will invoke it from the Java side where the UI resides. This work is
tracked by issue 1217186.

Change-Id: I24a18b8f41dcec0ff70ef07a840e996d2a3aff18
Bug: 1217185
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2953117
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Sophey <sophey@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892155}