Don't use a chrome extension origin in the DatabaseHelperTest
The DatabaseHelperTest.FetchData browser test tries to call
DatabaseTracker::DatabaseOpened with a chrome extension.  However,
chrome extensions are not part of the schema registry, which results in
the extension identifier being treated as an opaque origin.  This
problem manifests in
https://chromium-review.googlesource.com/c/chromium/src/+/2941645
because using an opaque origin causes the underlying storage APIs (e.g.
QuotaDatabase) to fail the DCHECK(!opaque()) when calling
StorageKey::Serialize.

This CL just changes the test to not use a chrome extension for an
identifier, as it is not relevant to what is being tested in
DatabaseHelperTest.

Bug: 1218217
Change-Id: I56ae4f2089b4cbccc2b68b5e3ca2be515e5cb465
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2951907
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Ali Beyad <abeyad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892162}