Reland: Refactor the -webkit-cross-fade() implementation
The current implementation largely live within the associated CSSValue.
This causes some duplication of code, and makes the code unnecessarily
complex when it comes to supporting different <image> arguments.

Instead move the implementation to a new StyleCrossfadeImage class,
keeping the argument <image>s as StyleImages. This makes it easier to
leverage the already existing StyleImage implementation to avoid
duplication of code.

ElementStyleResources needs to be taught how to perform the loading for
this new setup, so the existing functions are refactored into an
IsPending() query for checking if a CSSValue is pending 'load' and one
CachedStyleImage() for returning a cached StyleImage. The actual loading
is moved to a new StyleImageLoader helper class that traverses the
parameters ('parameter tree') and instantiates StyleImages.

This allows dropping some 'support' methods on CSSImageGeneratorValue,
such as IsFixed()/FixedSize()/LoadSubimages()/IsPending()/..., since
those only served to support the cross-fade implementation.

StylePendingImage::ComputedCSSValue() needs to be taught about the
potentially recursive <image> structure, so a helper class is added for
that as well.

Update: Make the IsEqual() implementation stricter (only match the same
object) to fix unregistration/removal of clients when the a new style is
computated that doesn't change the cross-fade() function.

Bug: 1175005
Change-Id: I0e9befa97c3bf17a964cde628e9a71e6ab0dc9c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2960451
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#892182}