Crash fix: TextFragmentSelectorGenerator
This is a speculative fix for a crash in TextFragmentSelectorGenerator.
This started after https://crrev.com/c/2927345 which changed the class
from keeping the selection as a Range to an RangeInFlatTree. Somehow
this lets through some degenerate cases of ranges.

I suspect these were issues prior to the above CL but were masked by
a conversion to Range which produced a valid selection but one that
wouldn't actually select the correct text.

Bug: 1216773
Change-Id: I25acc3fe23e295b6f0d5c06aff06c91158381fe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2956932
Reviewed-by: Gayane Petrosyan <gayane@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892208}