Fenced Frames: Correct the value of window.top for ShadowDOM-based impl
This CL implements a loose "fence" around the top-level <iframe> in a
<fencedframe>, particularly for `window.top` values. Specifically, this
CL restricts DOMWindow::top() to returning one of the following:
 - The nearest DOMWindow to the invoking-DOMWindow whose Owner() is
   non-null and whose Owner()'s FramePolicy's `is_fenced` bit is true,
   if such a DOMWindow exists.
 - Otherwise, return the normal DOMWindow that would be returned without
   this intervention.

The restriction described above is only enabled and relevant for the
FencedFrameShadowDOMDelegate implementation of fenced frames. The MPArch
implementation will give us this restriction by default, see [1].

In order to implement the above restrictions, this CL introduces a new
O(n) loop to DOMWindow::top(), where `n` is the height of the FrameTree.
This has no performance implications on the method however, because
before this CL, the method delegated to FrameTree::Top(), which itself
has an O(n) loop to climb the Frames in the FrameTree to the top.

The implications of this CL can be summarized like so:
 - `window.top` in the top-level internal <iframe> of a <fencedframe>
   element will always return `window`
 - `window.top` in subframes at any level underneath the top-level
   internal <iframe> in a <fencedframe> element will always return the
   `window` associated with the top-level internal <iframe>

Subsequent CLs will implement similar restrictions for other APIs (again
see [1]).

[1]: https://docs.google.com/document/d/1ijTZJT3DHQ1ljp4QQe4E4XCCRaYAxmInNzN1SzeJM8s/edit#heading=h.jztjmd6vstll

R=shivanisha@chromium.org

Bug: 1123606
Change-Id: I765037393f8d945fef8bd26d2b77302c6d7d67d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2945142
Commit-Queue: Dominic Farolino <dom@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892313}