Add the GetArcAppLaunchInfo interface.
ARC ghost window is launched during the system startup phase. When the
ARC app is ready, we need to get the ARC app launch parameters to
relaunch the ARC app, to replace the ghost window with the real ARC app
window.

This CL adds the GetArcAppLaunchInfo interface to get app launch info
from the full restore data, to relaunch the ARC app, when the ARC app is
ready.

BUG=1212584

Change-Id: I40b32b9fa8c909f9bc3a178f4a7eba10b09a8388
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2960329
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892337}