Don't show the restore setting notificaiton after reboot from crash.
The restore setting notification is shown when the restore button of the
full restore notification is clicked.

When the system reboots from a crash, whatever the full restore setting
is, the notification after a crash is always shown. So for the crash
restore notification, we don't need to show the restore setting
notification, which might confuse users.

Modify test cases to verify the restore setting notification is not
shown after clicking the restore button of the crash restore
notification.

BUG=1218397

Change-Id: Icadc8f7369cd434a2fd585a0c51b14f3e8b170d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2959248
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892345}